src/main/webapp/app/entities/abc-9/delete/abc-9-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Abc9Service } from '../service/abc-9.service';

import { Abc9DeleteDialogComponent } from './abc-9-delete-dialog.component';

describe('Abc9 Management Delete Component', () => {
  let comp: Abc9DeleteDialogComponent;
  let fixture: ComponentFixture<Abc9DeleteDialogComponent>;
  let service: Abc9Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [Abc9DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Abc9DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Abc9DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Abc9Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
